fix(winc): rename to wda and add version check
